Fixed /removexp command allowing xp removal below level one

+++ src/main/java/com/azuredoom/levelingcore/commands/RemoveXpCommand.java
@@ -63,6 +63,19 @@ public class RemoveXpCommand extends AbstractPlayerCommand {
         playerRef = this.playerArg.get(commandContext);
         var xpRef = this.xpArg.get(commandContext);
         var playerUUID = playerRef.getUuid();
+        var currentXp = levelService.getXp(playerUUID);
+        var totalXp = levelService.getXp(playerUUID);
+        var minXpForLevelOne = levelService.getXpForLevel(1);
+        if (totalXp - xpRef < minXpForLevelOne) {
+            commandContext.sendMessage(CommandLang.CANNOT_REMOVE_LEVEL_BELOW_ONE
+                    .param("player", playerRef.getUsername()));
+            return;
+        }
+        if (currentXp - xpRef < 0) {
+            commandContext.sendMessage(CommandLang.CANNOT_REMOVE_LEVEL_BELOW_ONE
+                    .param("player", playerRef.getUsername()));
+            return;
+        }
         levelService.removeXp(playerUUID, xpRef);
         var level = levelService.getLevel(playerUUID);
         var removedXPMsg = CommandLang.REMOVE_XP_1.param("xp", xpRef).param("player", playerRef.getUsername());
